feat: refactor server to fix load error
