+++ server/app.py
@@ -1,6 +1,5 @@
 from flask import Flask
 from handlers.route import configure_routes
-from clasification.model import MultinomialNaiveBayes, SVM, TFIDF
 from flask_cors import CORS
 
 app = Flask(__name__)
